product/requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
paramiko>=3.4.0
python-dotenv>=1.0.0
